Menu de Clientes
Adicionado o Menu de Clientes na aplicação de administrador, funcional e com a navegação entre telas implementada.

+++ E-commerce Application/src/SERVICE/ClienteSERVICE.java
@@ -22,6 +22,9 @@ public class ClienteSERVICE {
 
     public static boolean deletarCliente(String cpf) {
         ClienteMODEL clienteDeletar = ClienteDAO.selectClienteCpf(cpf);
+        if(clienteDeletar == null)
+            return false;
+        
         return ClienteDAO.deleteCliente(clienteDeletar);
     }
